+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 - Editor: Add support for uploading app icons.
 - Frontend: Implement error reporting using Sentry.
 
+### Changed
+
+- Frontend: Show the page title instead of the app title.
+
 ## [0.1.0] - 2018-10-19
 
 This is the initial release of the Appsemble platform.